@@ -1,8 +1,5 @@
 import re
 
-from lxml import html
-from lxml.html import HtmlComment
-
 
 class BasicBoxScoreRow:
     def __init__(self, html):
@@ -15,7 +12,7 @@ class BasicBoxScoreRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def minutes_played(self):
@@ -28,7 +25,7 @@ class BasicBoxScoreRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def attempted_field_goals(self):
@@ -37,7 +34,7 @@ class BasicBoxScoreRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def made_three_point_field_goals(self):
@@ -46,7 +43,7 @@ class BasicBoxScoreRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def attempted_three_point_field_goals(self):
@@ -55,7 +52,7 @@ class BasicBoxScoreRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def made_free_throws(self):
@@ -64,7 +61,7 @@ class BasicBoxScoreRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def attempted_free_throws(self):
@@ -73,7 +70,7 @@ class BasicBoxScoreRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def offensive_rebounds(self):
@@ -82,7 +79,7 @@ class BasicBoxScoreRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def defensive_rebounds(self):
@@ -91,7 +88,7 @@ class BasicBoxScoreRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def assists(self):
@@ -100,7 +97,7 @@ class BasicBoxScoreRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def steals(self):
@@ -109,7 +106,7 @@ class BasicBoxScoreRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def blocks(self):
@@ -118,7 +115,7 @@ class BasicBoxScoreRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def turnovers(self):
@@ -127,7 +124,7 @@ class BasicBoxScoreRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def personal_fouls(self):
@@ -136,7 +133,7 @@ class BasicBoxScoreRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def points(self):
@@ -145,7 +142,7 @@ class BasicBoxScoreRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def location_abbreviation(self):
@@ -154,8 +151,7 @@ class BasicBoxScoreRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
-
+        return ""
 
     @property
     def outcome(self):
@@ -164,7 +160,7 @@ class BasicBoxScoreRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def plus_minus(self):
@@ -173,7 +169,7 @@ class BasicBoxScoreRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def game_score(self):
@@ -182,7 +178,7 @@ class BasicBoxScoreRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
 
 class PlayerSeasonGameLogRow(BasicBoxScoreRow):
@@ -201,7 +197,7 @@ class PlayerSeasonGameLogRow(BasicBoxScoreRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def opponent_abbreviation(self):
@@ -210,7 +206,7 @@ class PlayerSeasonGameLogRow(BasicBoxScoreRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
 
 class PlayerBoxScoreRow(BasicBoxScoreRow):
@@ -229,7 +225,7 @@ class PlayerBoxScoreRow(BasicBoxScoreRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def opponent_abbreviation(self):
@@ -238,7 +234,7 @@ class PlayerBoxScoreRow(BasicBoxScoreRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
 
 class PlayerIdentificationRow:
@@ -258,15 +254,15 @@ class PlayerIdentificationRow:
     def slug(self):
         cell = self.player_cell
         if cell is None:
-            return ''
+            return ""
 
-        return cell.get('data-append-csv')
+        return cell.get("data-append-csv")
 
     @property
     def name(self):
         cell = self.player_cell
         if cell is None:
-            return ''
+            return ""
 
         return cell.text_content()
 
@@ -284,7 +280,7 @@ class PlayerAdvancedSeasonTotalsTable:
             /tbody
             /tr[
                 (
-                    not(contains(@class, 'thead')) and 
+                    not(contains(@class, 'thead')) and
                     not(contains(@class, 'norank'))
                 )
             ]
@@ -294,7 +290,9 @@ class PlayerAdvancedSeasonTotalsTable:
         player_advanced_season_totals_rows = []
         for row_html in self.html.xpath(self.rows_query):
             row = PlayerAdvancedSeasonTotalsRow(html=row_html)
-            if (include_combined_totals is True and row.is_combined_totals is True) or row.is_combined_totals is False:
+            if (
+                include_combined_totals is True and row.is_combined_totals is True
+            ) or row.is_combined_totals is False:
                 # Basketball Reference includes a "total" row for players that got traded
                 # which is essentially a sum of all player team rows
                 # I want to avoid including those, so I check the "team" field value for "TOT"
@@ -315,7 +313,7 @@ class PlayerSeasonTotalTable:
                     //table[@id='totals_stats']
                     /tbody
                     /tr[
-                        not(contains(@class, 'thead')) and 
+                        not(contains(@class, 'thead')) and
                         not(contains(@class, 'norank'))
                     ]
                 """
@@ -351,15 +349,15 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
     def slug(self):
         cell = self.player_cell
         if cell is None:
-            return ''
+            return ""
 
-        return cell.get('data-append-csv')
+        return cell.get("data-append-csv")
 
     @property
     def name(self):
         cell = self.player_cell
         if cell is None:
-            return ''
+            return ""
 
         return cell.text_content()
 
@@ -370,7 +368,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def age(self):
@@ -379,7 +377,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def team_abbreviation(self):
@@ -388,7 +386,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def games_played(self):
@@ -397,7 +395,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def minutes_played(self):
@@ -406,7 +404,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def player_efficiency_rating(self):
@@ -415,7 +413,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def true_shooting_percentage(self):
@@ -424,7 +422,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def three_point_attempt_rate(self):
@@ -433,7 +431,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def free_throw_attempt_rate(self):
@@ -442,7 +440,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def offensive_rebound_percentage(self):
@@ -451,7 +449,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def defensive_rebound_percentage(self):
@@ -460,7 +458,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def total_rebound_percentage(self):
@@ -469,7 +467,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def assist_percentage(self):
@@ -478,7 +476,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def steal_percentage(self):
@@ -487,7 +485,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def block_percentage(self):
@@ -496,7 +494,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def turnover_percentage(self):
@@ -505,7 +503,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def usage_percentage(self):
@@ -514,7 +512,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def offensive_win_shares(self):
@@ -523,7 +521,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def defensive_win_shares(self):
@@ -532,7 +530,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def win_shares(self):
@@ -541,7 +539,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def win_shares_per_48_minutes(self):
@@ -550,7 +548,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def offensive_plus_minus(self):
@@ -559,7 +557,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def defensive_plus_minus(self):
@@ -568,7 +566,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def plus_minus(self):
@@ -577,7 +575,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def value_over_replacement_player(self):
@@ -586,7 +584,7 @@ class PlayerAdvancedSeasonTotalsRow(PlayerIdentificationRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def is_combined_totals(self):
@@ -607,7 +605,7 @@ class PlayerSeasonTotalsRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def age(self):
@@ -616,7 +614,7 @@ class PlayerSeasonTotalsRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def games_played(self):
@@ -625,7 +623,7 @@ class PlayerSeasonTotalsRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def games_started(self):
@@ -634,7 +632,7 @@ class PlayerSeasonTotalsRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def is_combined_totals(self):
@@ -650,7 +648,7 @@ class PlayerSeasonTotalsRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def player_cell(self):
@@ -665,15 +663,15 @@ class PlayerSeasonTotalsRow:
     def slug(self):
         cell = self.player_cell
         if cell is None:
-            return ''
+            return ""
 
-        return cell.get('data-append-csv')
+        return cell.get("data-append-csv")
 
     @property
     def name(self):
         cell = self.player_cell
         if cell is None:
-            return ''
+            return ""
 
         return cell.text_content()
 
@@ -684,7 +682,7 @@ class PlayerSeasonTotalsRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def minutes_played(self):
@@ -697,7 +695,7 @@ class PlayerSeasonTotalsRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def attempted_field_goals(self):
@@ -706,7 +704,7 @@ class PlayerSeasonTotalsRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def made_three_point_field_goals(self):
@@ -715,7 +713,7 @@ class PlayerSeasonTotalsRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def attempted_three_point_field_goals(self):
@@ -724,7 +722,7 @@ class PlayerSeasonTotalsRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def made_free_throws(self):
@@ -733,7 +731,7 @@ class PlayerSeasonTotalsRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def attempted_free_throws(self):
@@ -742,7 +740,7 @@ class PlayerSeasonTotalsRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def offensive_rebounds(self):
@@ -751,7 +749,7 @@ class PlayerSeasonTotalsRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def defensive_rebounds(self):
@@ -760,7 +758,7 @@ class PlayerSeasonTotalsRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def assists(self):
@@ -769,7 +767,7 @@ class PlayerSeasonTotalsRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def steals(self):
@@ -778,7 +776,7 @@ class PlayerSeasonTotalsRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def blocks(self):
@@ -787,7 +785,7 @@ class PlayerSeasonTotalsRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def turnovers(self):
@@ -796,7 +794,7 @@ class PlayerSeasonTotalsRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def personal_fouls(self):
@@ -805,7 +803,7 @@ class PlayerSeasonTotalsRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def points(self):
@@ -814,7 +812,7 @@ class PlayerSeasonTotalsRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
 
 class BoxScoresPage:
@@ -825,7 +823,9 @@ class BoxScoresPage:
     def statistics_tables(self):
         return [
             StatisticsTable(table_html)
-            for table_html in self.html.xpath('//table[contains(@class, "stats_table")]')
+            for table_html in self.html.xpath(
+                '//table[contains(@class, "stats_table")]'
+            )
         ]
 
     @property
@@ -843,18 +843,18 @@ class StatisticsTable:
 
     @property
     def has_basic_statistics(self):
-        return 'game-basic' in self.html.attrib["id"]
+        return "game-basic" in self.html.attrib["id"]
 
     @property
     def team_abbreviation(self):
         # Example id value is box-BOS-game-basic or box-BOS-game-advanced
-        match = re.match('^box-(.+)-game', self.html.attrib["id"])
+        match = re.match("^box-(.+)-game", self.html.attrib["id"])
         return match.group(1)
 
     @property
     def team_totals(self):
         # Team totals are stored as table footers
-        footers = self.html.xpath('tfoot/tr')
+        footers = self.html.xpath("tfoot/tr")
         if len(footers) > 0:
             return BasicBoxScoreRow(html=footers[0])
 
@@ -869,7 +869,9 @@ class DailyLeadersPage:
     def daily_leaders(self):
         return [
             PlayerGameBoxScoreRow(row_html)
-            for row_html in self.html.xpath('//table[@id="stats"]//tbody/tr[not(contains(@class, "thead"))]')
+            for row_html in self.html.xpath(
+                '//table[@id="stats"]//tbody/tr[not(contains(@class, "thead"))]'
+            )
         ]
 
 
@@ -937,7 +939,7 @@ class PlayerSeasonBoxScoresRow(PlayerSeasonGameLogRow):
         # have a "colspan" attribute
         cells = self.html.xpath('td[@data-stat="is_starter"]')
         if len(cells) > 0:
-            colspan_value = cells[0].get('colspan', None)
+            colspan_value = cells[0].get("colspan", None)
             return colspan_value is None
 
         return False
@@ -949,7 +951,7 @@ class PlayerSeasonBoxScoresRow(PlayerSeasonGameLogRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def points_scored(self):
@@ -958,7 +960,7 @@ class PlayerSeasonBoxScoresRow(PlayerSeasonGameLogRow):
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
 
 class PlayerGameBoxScoreRow(PlayerBoxScoreRow, PlayerIdentificationRow):
@@ -976,11 +978,7 @@ class PlayByPlayPage:
 
     @property
     def team_names_query(self):
-        return \
-            '//*[@id="content"]' \
-            '//div[@class="scorebox"]' \
-            '//strong' \
-            '//a'
+        return '//*[@id="content"]//div[@class="scorebox"]//strong//a'
 
     @property
     def play_by_play_table(self):
@@ -990,10 +988,7 @@ class PlayByPlayPage:
     def team_names(self):
         names = self.html.xpath(self.team_names_query)
 
-        return [
-            name.text_content()
-            for name in names
-        ]
+        return [name.text_content() for name in names]
 
     @property
     def away_team_name(self):
@@ -1031,17 +1026,17 @@ class PlayByPlayRow:
 
     @property
     def away_team_play_description(self):
-        if 6 == len(self.html):
+        if len(self.html) == 6:
             return self.html[1].text_content().strip()
 
-        return ''
+        return ""
 
     @property
     def home_team_play_description(self):
-        if 6 == len(self.html):
+        if len(self.html) == 6:
             return self.html[5].text_content().strip()
 
-        return ''
+        return ""
 
     @property
     def is_away_team_play(self):
@@ -1053,26 +1048,28 @@ class PlayByPlayRow:
 
     @property
     def formatted_scores(self):
-        if 6 == len(self.html):
+        if len(self.html) == 6:
             return self.html[3].text_content().strip()
-        return ''
+        return ""
 
     @property
     def is_start_of_period(self):
-        return self.timestamp_cell.get('colspan') == '6'
+        return self.timestamp_cell.get("colspan") == "6"
 
     @property
     def has_play_by_play_data(self):
-        # TODO: @jaebradley refactor this to be slightly clearer
+        # TODO: @nickth3man refactor this to be slightly clearer
         # Need to avoid rows that indicate start of period
         # Or denote tipoff / end of period (colspan = 5)
         # Or are one of the table headers for each period group (aria-label = Time)
         # There are certain cases, like at the 10 minute mark in https://www.basketball-reference.com/boxscores/pbp/199911160ATL.html
         # where there are no event details. Probably a visual bug on Basketball Reference's side of things.
-        return not self.is_start_of_period \
-            and 2 <= len(self.html) \
-            and self.html[1].get('colspan') != '5' \
-            and self.timestamp_cell.get('aria-label') != 'Time'
+        return (
+            not self.is_start_of_period
+            and len(self.html) >= 2
+            and self.html[1].get("colspan") != "5"
+            and self.timestamp_cell.get("aria-label") != "Time"
+        )
 
 
 class DailyBoxScoresPage:
@@ -1086,7 +1083,7 @@ class DailyBoxScoresPage:
     @property
     def game_url_paths(self):
         game_links = self.html.xpath(self.game_url_paths_query)
-        return [game_link.attrib['href'] for game_link in game_links]
+        return [game_link.attrib["href"] for game_link in game_links]
 
 
 class SchedulePage:
@@ -1095,10 +1092,12 @@ class SchedulePage:
 
     @property
     def other_months_schedule_links_query(self):
-        return '//div[@id="content"]' \
-               '/div[@class="filter"]' \
-               '/div[not(contains(@class, "current"))]' \
-               '/a'
+        return (
+            '//div[@id="content"]'
+            '/div[@class="filter"]'
+            '/div[not(contains(@class, "current"))]'
+            "/a"
+        )
 
     @property
     def rows_query(self):
@@ -1107,10 +1106,7 @@ class SchedulePage:
     @property
     def other_months_schedule_urls(self):
         links = self.html.xpath(self.other_months_schedule_links_query)
-        return [
-            link.attrib['href']
-            for link in links
-        ]
+        return [link.attrib["href"] for link in links]
 
     @property
     def rows(self):
@@ -1119,7 +1115,7 @@ class SchedulePage:
             for row in self.html.xpath(self.rows_query)
             # Every row in each month's schedule table represents a game
             # except for the row where the only content is "Playoffs"
-            if row.text_content() != 'Playoffs'
+            if row.text_content() != "Playoffs"
         ]
 
 
@@ -1139,7 +1135,7 @@ class ScheduleRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def start_time_of_day(self):
@@ -1148,7 +1144,7 @@ class ScheduleRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def away_team_name(self):
@@ -1157,7 +1153,7 @@ class ScheduleRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def home_team_name(self):
@@ -1166,7 +1162,7 @@ class ScheduleRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def away_team_score(self):
@@ -1175,7 +1171,7 @@ class ScheduleRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
     @property
     def home_team_score(self):
@@ -1184,7 +1180,7 @@ class ScheduleRow:
         if len(cells) > 0:
             return cells[0].text_content()
 
-        return ''
+        return ""
 
 
 class SearchPage:
@@ -1197,15 +1193,13 @@ class SearchPage:
 
     @property
     def nba_aba_baa_players_pagination_links_query(self):
-        return '{NBA_ABA_PLAYERS_CONTENT_QUERY}/div[@class="search-pagination"]/a'.format(
-            NBA_ABA_PLAYERS_CONTENT_QUERY=self.nba_aba_baa_players_content_query
+        return (
+            f'{self.nba_aba_baa_players_content_query}/div[@class="search-pagination"]/a'
         )
 
     @property
     def nba_aba_baa_player_search_items_query(self):
-        return '{NBA_ABA_PLAYERS_CONTENT_QUERY}/div[@class="search-item"]'.format(
-            NBA_ABA_PLAYERS_CONTENT_QUERY=self.nba_aba_baa_players_content_query
-        )
+        return f'{self.nba_aba_baa_players_content_query}/div[@class="search-item"]'
 
     @property
     def nba_aba_baa_players_pagination_links(self):
@@ -1221,7 +1215,7 @@ class SearchPage:
         first_link = links[0]
 
         if len(links) == 1:
-            if first_link.text_content() == 'Previous 100 Results':
+            if first_link.text_content() == "Previous 100 Results":
                 return None
 
             return first_link.attrib["href"]
@@ -1232,7 +1226,9 @@ class SearchPage:
     def nba_aba_baa_players(self):
         return [
             PlayerSearchResult(html=result_html)
-            for result_html in self.html.xpath(self.nba_aba_baa_player_search_items_query)
+            for result_html in self.html.xpath(
+                self.nba_aba_baa_player_search_items_query
+            )
         ]
 
 
@@ -1319,7 +1315,7 @@ class PlayerPageTotalsTable:
     def rows(self):
         return [
             PlayerPageTotalsRow(html=row_html)
-            for row_html in self.html.xpath('.//tbody/tr')
+            for row_html in self.html.xpath(".//tbody/tr")
         ]
 
     def __eq__(self, other):
@@ -1396,7 +1392,7 @@ class ConferenceDivisionStandingsTable:
     def rows(self):
         return [
             ConferenceDivisionStandingsRow(html=row_html)
-            for row_html in self.html.xpath('.//tbody/tr')
+            for row_html in self.html.xpath(".//tbody/tr")
         ]
 
 
@@ -1406,15 +1402,15 @@ class ConferenceDivisionStandingsRow:
 
     @property
     def is_division_name_row(self):
-        return self.html.attrib['class'] == 'thead'
+        return self.html.attrib["class"] == "thead"
 
     @property
     def is_standings_row(self):
-        return self.html.attrib['class'] == 'full_table'
+        return self.html.attrib["class"] == "full_table"
 
     @property
     def division_name(self):
-        cells = self.html.xpath('.//th')
+        cells = self.html.xpath(".//th")
 
         if len(cells) == 1:
             return cells[0].text_content()
@@ -1457,15 +1453,15 @@ class PlayerContractsRow:
     def player_name(self):
         matching_cells = self.html.xpath('.//td[@data-stat="player"]')
 
-        if 1 == len(matching_cells):
+        if len(matching_cells) == 1:
             return matching_cells[0].text_content()
 
         return None
 
     @property
     def player_identifier(self):
-        matching_attribute_value = self.html.xpath('.//td/@data-append-csv')
-        if 1 == len(matching_attribute_value):
+        matching_attribute_value = self.html.xpath(".//td/@data-append-csv")
+        if len(matching_attribute_value) == 1:
             return matching_attribute_value[0]
 
         return None
@@ -1474,50 +1470,64 @@ class PlayerContractsRow:
     def team_abbreviation(self):
         matching_cells = self.html.xpath('.//td[@data-stat="team_id"]')
 
-        if 1 == len(matching_cells):
+        if len(matching_cells) == 1:
             return matching_cells[0].text_content()
 
         return None
 
     @property
     def first_contract_year_data(self):
-        return self.calculate_contract_year_data(contract_year_data_stat_attribute_value="y1")
+        return self.calculate_contract_year_data(
+            contract_year_data_stat_attribute_value="y1"
+        )
 
     @property
     def second_contract_year_data(self):
-        return self.calculate_contract_year_data(contract_year_data_stat_attribute_value="y2")
+        return self.calculate_contract_year_data(
+            contract_year_data_stat_attribute_value="y2"
+        )
 
     @property
     def third_contract_year_data(self):
-        return self.calculate_contract_year_data(contract_year_data_stat_attribute_value="y3")
+        return self.calculate_contract_year_data(
+            contract_year_data_stat_attribute_value="y3"
+        )
 
     @property
     def fourth_contract_year_data(self):
-        return self.calculate_contract_year_data(contract_year_data_stat_attribute_value="y4")
+        return self.calculate_contract_year_data(
+            contract_year_data_stat_attribute_value="y4"
+        )
 
     @property
     def fifth_contract_year_data(self):
-        return self.calculate_contract_year_data(contract_year_data_stat_attribute_value="y5")
+        return self.calculate_contract_year_data(
+            contract_year_data_stat_attribute_value="y5"
+        )
 
     @property
     def sixth_contract_year_data(self):
-        return self.calculate_contract_year_data(contract_year_data_stat_attribute_value="y6")
+        return self.calculate_contract_year_data(
+            contract_year_data_stat_attribute_value="y6"
+        )
 
     @property
     def guaranteed(self):
         matching_cells = self.html.xpath('.//td[@data-stat="remain_gtd"]')
 
-        if 1 == len(matching_cells):
+        if len(matching_cells) == 1:
             return matching_cells[0].text_content()
 
         return None
 
     def calculate_contract_year_data(self, contract_year_data_stat_attribute_value):
-        matching_cells = self.html.xpath(f'.//td[@data-stat="{contract_year_data_stat_attribute_value}"]')
+        matching_cells = self.html.xpath(
+            f'.//td[@data-stat="{contract_year_data_stat_attribute_value}"]'
+        )
 
-        if 1 == len(matching_cells):
+        if len(matching_cells) == 1:
             salary = matching_cells[0].text_content()
-            class_names = matching_cells[0].get('class')
+            class_names = matching_cells[0].get("class")
 
             return salary, class_names
 
