@@ -1,11 +1,21 @@
 import requests
 from lxml import html
 
-from src.data import TEAM_TO_TEAM_ABBREVIATION, TeamTotal, PlayerData
+from src.data import TEAM_TO_TEAM_ABBREVIATION, PlayerData, TeamTotal
 from src.errors import InvalidDate, InvalidPlayerAndSeason
-from src.html import DailyLeadersPage, PlayerSeasonBoxScoresPage, PlayerSeasonTotalTable, \
-    PlayerAdvancedSeasonTotalsTable, PlayByPlayPage, SchedulePage, BoxScoresPage, DailyBoxScoresPage, SearchPage, \
-    PlayerPage, StandingsPage
+from src.html import (
+    BoxScoresPage,
+    DailyBoxScoresPage,
+    DailyLeadersPage,
+    PlayByPlayPage,
+    PlayerAdvancedSeasonTotalsTable,
+    PlayerPage,
+    PlayerSeasonBoxScoresPage,
+    PlayerSeasonTotalTable,
+    SchedulePage,
+    SearchPage,
+    StandingsPage,
+)
 
 
 class HTTPService:
@@ -15,10 +25,7 @@ class HTTPService:
         self.parser = parser
 
     def standings(self, season_end_year):
-        url = '{BASE_URL}/leagues/NBA_{season_end_year}.html'.format(
-            BASE_URL=HTTPService.BASE_URL,
-            season_end_year=season_end_year,
-        )
+        url = f'{HTTPService.BASE_URL}/leagues/NBA_{season_end_year}.html'
 
         response = requests.get(url=url, allow_redirects=False)
 
@@ -29,12 +36,7 @@ class HTTPService:
                self.parser.parse_division_standings(standings=page.division_standings.western_conference_table.rows)
 
     def player_box_scores(self, day, month, year):
-        url = '{BASE_URL}/friv/dailyleaders.cgi?month={month}&day={day}&year={year}'.format(
-            BASE_URL=HTTPService.BASE_URL,
-            day=day,
-            month=month,
-            year=year
-        )
+        url = f'{HTTPService.BASE_URL}/friv/dailyleaders.cgi?month={month}&day={day}&year={year}'
 
         response = requests.get(url=url, allow_redirects=False)
 
@@ -50,13 +52,7 @@ class HTTPService:
         # Makes assumption that basketball reference pattern of breaking out player pathing using first character of
         # surname can be derived from the fact that basketball reference also has a pattern of player identifiers
         # starting with first few characters of player's surname
-        url = '{BASE_URL}/players/{player_surname_starting_character}/{player_identifier}/gamelog/{season_end_year}' \
-            .format(
-            BASE_URL=HTTPService.BASE_URL,
-            player_surname_starting_character=player_identifier[0],
-            player_identifier=player_identifier,
-            season_end_year=season_end_year,
-        )
+        url = f'{HTTPService.BASE_URL}/players/{player_identifier[0]}/{player_identifier}/gamelog/{season_end_year}'
 
         response = requests.get(url=url, allow_redirects=False)
         response.raise_for_status()
@@ -71,13 +67,7 @@ class HTTPService:
         # Makes assumption that basketball reference pattern of breaking out player pathing using first character of
         # surname can be derived from the fact that basketball reference also has a pattern of player identifiers
         # starting with first few characters of player's surname
-        url = '{BASE_URL}/players/{player_surname_starting_character}/{player_identifier}/gamelog/{season_end_year}' \
-            .format(
-            BASE_URL=HTTPService.BASE_URL,
-            player_surname_starting_character=player_identifier[0],
-            player_identifier=player_identifier,
-            season_end_year=season_end_year,
-        )
+        url = f'{HTTPService.BASE_URL}/players/{player_identifier[0]}/{player_identifier}/gamelog/{season_end_year}'
 
         response = requests.get(url=url, allow_redirects=False)
         response.raise_for_status()
@@ -89,13 +79,11 @@ class HTTPService:
         return self.parser.parse_player_season_box_scores(box_scores=page.playoff_box_scores_table.rows, include_inactive_games=include_inactive_games)
 
     def play_by_play(self, home_team, day, month, year):
-        add_0_if_needed = lambda s: "0" + s if len(s) == 1 else s
+        def add_0_if_needed(s):
+            return "0" + s if len(s) == 1 else s
 
         # the hard-coded `0` in the url assumes we always take the first match of the given date and team.
-        url = "{BASE_URL}/boxscores/pbp/{year}{month}{day}0{team_abbr}.html".format(
-            BASE_URL=HTTPService.BASE_URL, year=year, month=add_0_if_needed(str(month)), day=add_0_if_needed(str(day)),
-            team_abbr=TEAM_TO_TEAM_ABBREVIATION[home_team]
-        )
+        url = f"{HTTPService.BASE_URL}/boxscores/pbp/{year}{add_0_if_needed(str(month))}{add_0_if_needed(str(day))}0{TEAM_TO_TEAM_ABBREVIATION[home_team]}.html"
         response = requests.get(url=url)
         response.raise_for_status()
 
@@ -108,10 +96,7 @@ class HTTPService:
         )
 
     def players_advanced_season_totals(self, season_end_year, include_combined_values=False):
-        url = '{BASE_URL}/leagues/NBA_{season_end_year}_advanced.html'.format(
-            BASE_URL=HTTPService.BASE_URL,
-            season_end_year=season_end_year,
-        )
+        url = f'{HTTPService.BASE_URL}/leagues/NBA_{season_end_year}_advanced.html'
 
         response = requests.get(url=url)
 
@@ -121,10 +106,7 @@ class HTTPService:
         return self.parser.parse_player_advanced_season_totals_parser(totals=table.get_rows(include_combined_values))
 
     def players_season_totals(self, season_end_year):
-        url = '{BASE_URL}/leagues/NBA_{season_end_year}_totals.html'.format(
-            BASE_URL=HTTPService.BASE_URL,
-            season_end_year=season_end_year,
-        )
+        url = f'{HTTPService.BASE_URL}/leagues/NBA_{season_end_year}_totals.html'
 
         response = requests.get(url=url)
 
@@ -142,10 +124,7 @@ class HTTPService:
         return self.parser.parse_scheduled_games(games=page.rows)
 
     def season_schedule(self, season_end_year):
-        url = '{BASE_URL}/leagues/NBA_{season_end_year}_games.html'.format(
-            BASE_URL=HTTPService.BASE_URL,
-            season_end_year=season_end_year
-        )
+        url = f'{HTTPService.BASE_URL}/leagues/NBA_{season_end_year}_games.html'
 
         response = requests.get(url=url)
 
@@ -155,14 +134,14 @@ class HTTPService:
         season_schedule_values = self.parser.parse_scheduled_games(games=page.rows)
 
         for month_url_path in page.other_months_schedule_urls:
-            url = '{BASE_URL}{month_url_path}'.format(BASE_URL=HTTPService.BASE_URL, month_url_path=month_url_path)
+            url = f'{HTTPService.BASE_URL}{month_url_path}'
             monthly_schedule = self.schedule_for_month(url=url)
             season_schedule_values.extend(monthly_schedule)
 
         return season_schedule_values
 
     def team_box_score(self, game_url_path):
-        url = "{BASE_URL}/{game_url_path}".format(BASE_URL=HTTPService.BASE_URL, game_url_path=game_url_path)
+        url = f"{HTTPService.BASE_URL}/{game_url_path}"
 
         response = requests.get(url=url)
 
@@ -180,7 +159,7 @@ class HTTPService:
         )
 
     def team_box_scores(self, day, month, year):
-        url = "{BASE_URL}/boxscores/".format(BASE_URL=HTTPService.BASE_URL)
+        url = f"{HTTPService.BASE_URL}/boxscores/"
 
         response = requests.get(url=url, params={"day": day, "month": month, "year": year})
 
@@ -196,7 +175,7 @@ class HTTPService:
 
     def search(self, term):
         response = requests.get(
-            url="{BASE_URL}/search/search.fcgi".format(BASE_URL=HTTPService.BASE_URL),
+            url=f"{HTTPService.BASE_URL}/search/search.fcgi",
             params={"search": term}
         )
 
@@ -204,17 +183,14 @@ class HTTPService:
 
         player_results = []
 
-        if response.url.startswith("{BASE_URL}/search/search.fcgi".format(BASE_URL=HTTPService.BASE_URL)):
+        if response.url.startswith(f"{HTTPService.BASE_URL}/search/search.fcgi"):
             page = SearchPage(html=html.fromstring(response.content))
             parsed_results = self.parser.parse_player_search_results(nba_aba_baa_players=page.nba_aba_baa_players)
             player_results += parsed_results["players"]
 
             while page.nba_aba_baa_players_pagination_url is not None:
                 response = requests.get(
-                    url="{BASE_URL}/search/{pagination_url}".format(
-                        BASE_URL=HTTPService.BASE_URL,
-                        pagination_url=page.nba_aba_baa_players_pagination_url
-                    )
+                    url=f"{HTTPService.BASE_URL}/search/{page.nba_aba_baa_players_pagination_url}"
                 )
 
                 response.raise_for_status()
@@ -224,16 +200,16 @@ class HTTPService:
                 parsed_results = self.parser.parse_player_search_results(nba_aba_baa_players=page.nba_aba_baa_players)
                 player_results += parsed_results["players"]
 
-        elif response.url.startswith("{BASE_URL}/players".format(BASE_URL=HTTPService.BASE_URL)):
+        elif response.url.startswith(f"{HTTPService.BASE_URL}/players"):
             page = PlayerPage(html=html.fromstring(response.content))
             data = PlayerData(
                 name=page.name,
                 resource_location=response.url,
-                league_abbreviations=set([
+                league_abbreviations={
                     row.league_abbreviation
                     for row in page.totals_table.rows
                     if row.league_abbreviation is not None
-                ])
+                }
             )
             player_results += [self.parser.parse_player_data(player=data)]
 
